// generated/src/app/api-models/report-selection-param.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ReportSelectionParam {
  seqSelect: number;
  seqJobId: number;
  selectionTypeCode: string;
  columnName: string;
  columnOccurrence: number;
  operator: string;
  fromValue: string;
  thruValue: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  columnType: string;
}